requirements.txt
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
lightgbm>=3.3
catboost>=1.0
matplotlib>=3.4
seaborn>=0.11
joblib
pyyaml
pytest
shap
